streamlit/requirements.txt
streamlit==1.45.1
psycopg2-binary==2.9.10
pandas==2.2.3
plotly==6.1.0
